backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
eventlet==0.35.2
flask-sqlalchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
boto3==1.34.0
lasio==0.32
python-dotenv==1.0.0
groq==0.4.2
Werkzeug==3.0.1
